Rename postgres utilities
